% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNextGeneration.R
\name{create_next_generation}
\alias{create_next_generation}
\title{create_next_generation function}
\usage{
create_next_generation(generation_t0, objFunOutput_t0, iter)
}
\arguments{
\item{generation_t0, }{objFunOutput_t0, iter}
}
\description{
This function creates new generations of model
}
\examples{
create_next_generation(generation_t0, objFunOutput_t0, iter)
}
\keyword{generation}
\keyword{next}
