% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateFitness.R
\name{evaluate_fitness}
\alias{evaluate_fitness}
\title{evaluate_fitness function}
\usage{
evaluate_fitness(generation_t0, Y, X)
}
\arguments{
\item{genereation_t0, }{Y, X}
}
\description{
This function evaluates the goodness of fit with user specified criteria
}
\examples{
function(generation_t0, Y, X)
}
\keyword{fitness}
