% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFounders.R
\name{generate_founders}
\alias{generate_founders}
\title{generate_founders function}
\usage{
generate_founders(X)
}
\arguments{
\item{x(data}{set)}
}
\description{
This function generates the first generation of models.
}
\examples{
generate_founders(mtcars)
}
\keyword{founders}
