% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{select}
\alias{select}
\title{select function}
\usage{
select(Y, X, iter, objFun = c("AIC", "BIC", "logLik", "user"),
  family = "gaussian", crossMeth = c("method1", "method2", "method3"),
  numChromosomes = NULL, pCrossover = 1, user_objFun = NULL,
  converge = TRUE, minimize = TRUE, parallel = FALSE)
}
\arguments{
\item{Y, }{X, iter, objFun = c("AIC", "BIC", "logLik", "user"),
family = "gaussian",
crossMeth = c("method1", "method2", "method3"),
numChromosomes = NULL,
pCrossover = 1,
user_objFun = NULL,
converge = TRUE, minimize = TRUE, parallel = FALSE}
}
\description{
This function is the primary function that do variable selection.
}
\examples{
select()
}
\keyword{cats}
